+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2014-09-11 (0.1.0)
+
+* Added support for node-mysql 2.5.0 enqueue event
+* Deprecated connection 'queue' event
+
 ## 2014-08-09 (0.0.14)
 
 * Initial commit